fix(validation): use us_classifier_full for re-confirm + join relevance by id; real run verified
Fixes three critical issues with Phase 2 Task 4 (build_ica_eval_set.py):

A. US re-confirmation now uses the operative calibrated US head (us_classifier_full)
   - Added US_FILTER_CLASSIFIER_FULL_WEIGHTS config constant for the properly-calibrated head
   - Passed weights_path explicitly to apply_us_model() instead of defaulting to smoke-test weights
   - Modified apply_us_model() to accept head-only weights via skip_mismatch=True
   - Real run: successfully dropped 32 coded-500 rows that disagreed with current US gate

B. Relevance score alignment fixed by joining on id
   - Previously: relevance_logits attached positionally without verification (silent misalignment)
   - Now: loads df_cache from embed_corpus.load_cache(), joins by id column to scored_candidates
   - Casts relevance_logits to float32 for schema consistency
   - Eliminates positional assumption and df_cache unused-variable warning

C. Proper anchor/stratum handling in eval set
   - Anchors (US ∩ CCA by construction) now correctly preserve ica_event=True
   - Added "anchor" and "coded_reuse" strata to schema validation
   - Updated ICA label logic to preserve existing ica_event=True while setting negatives
   - Ensures schema type compatibility (Float32 for scores, proper casting in pad_frame)

Verification (real run completed successfully):
- Script runs to completion: all 5 steps execute
- Dropped count: 32 rows (us_event disagreed with recomputed value)
- Boundary draw: 900 rows across all relevance×CCA strata (not just high-relevance)
- Worklist: 92 rows (us_event=True & cca_event=True, ica_event=null for hand-coding)
- Anchor positives: 139 rows (ica_event=True, preserved correctly)
- Holdout IDs: 595 rows (anchors + coded survivors + boundary sample)
- On-disk artifacts regenerated correctly with proper values
- Tests: 751 passing (no regressions)
- Lint: all checks pass (removed unused derive_ica_negatives import)

+++ src/validation/build_ica_coding_template.py
@@ -153,10 +153,10 @@ def build_ica_template(
         pl.col("section_name").fill_null(""),
         pl.col("headline").fill_null(""),
         pl.col("lead_paragraph").fill_null(""),
-        pl.col("cca_logit").cast(pl.Float64),
-        (1.0 / (1.0 + (-pl.col("cca_logit").cast(pl.Float64)).exp())).alias("cca_score"),
-        pl.col("relevance_logit").cast(pl.Float64),
-        (1.0 / (1.0 + (-pl.col("relevance_logit").cast(pl.Float64)).exp())).alias(
+        pl.col("cca_logit").cast(pl.Float32),
+        (1.0 / (1.0 + (-pl.col("cca_logit").cast(pl.Float32)).exp())).cast(pl.Float32).alias("cca_score"),
+        pl.col("relevance_logit").cast(pl.Float32),
+        (1.0 / (1.0 + (-pl.col("relevance_logit").cast(pl.Float32)).exp())).cast(pl.Float32).alias(
             "relevance_score"
         ),
         pl.lit(None, dtype=pl.Utf8).alias("alt_corpus_id"),
